Rename a certain promotion
